feat(richtext-lexical)!: finalize ClientFeature interface
